feat: pedindo confirmacao do user para apagr opiniao / inicio API game cover

+++ src/main/resources/static/js/opiniao/deleteOpiniao.js
@@ -9,31 +9,37 @@ document.addEventListener('DOMContentLoaded', function() {
         const [name, value] = cookie.split('=');
         if (name === 'opinion_id') {
             opinionId = value;
-                break;
+            break;
         }
     }
-    
+
     button.addEventListener('click', function() {
-        fetch(`http://localhost:8080/opinions/delete/${opinionId}`, {
-            method: 'DELETE',
-            headers: {
-                Authorization: `Bearer ${token}`
-            }
-        }).then(response => {
-            if (!response.ok) {
-                if (response.status === 400) {
-                    return response.text().then(errorMessage => {
-                        alert('Erro 400: ' + errorMessage);
-                        throw new Error(`Erro ${response.status}: ${errorMessage}`);
-                    });
-                } else if (response.status === 401 || response.status === 403) {
-                    throw new Error(`Erro ${response.status}: Você não está autorizado para a operação desejada`);
-                } else {
-                    alert('Erro ao excluir opinião. Por favor, tente novamente.');
+        const userConfirmed = window.confirm('Você tem certeza que deseja excluir esta opinião? Esta ação não poderá ser desfeita.');
+
+        if (userConfirmed) {
+            fetch(`http://localhost:8080/opinions/delete/${opinionId}`, {
+                method: 'DELETE',
+                headers: {
+                    Authorization: `Bearer ${token}`
                 }
-            }
-            alert('Opinião excluída com sucesso!');
-            window.location.href = 'http://localhost:1313/opinioes';
-        });
+            }).then(response => {
+                if (!response.ok) {
+                    if (response.status === 400) {
+                        return response.text().then(errorMessage => {
+                            alert('Erro 400: ' + errorMessage);
+                            throw new Error(`Erro ${response.status}: ${errorMessage}`);
+                        });
+                    } else if (response.status === 401 || response.status === 403) {
+                        throw new Error(`Erro ${response.status}: Você não está autorizado para a operação desejada`);
+                    } else {
+                        alert('Erro ao excluir opinião. Por favor, tente novamente.');
+                    }
+                }
+                alert('Opinião excluída com sucesso!');
+                window.location.href = 'http://localhost:1313/opinioes';
+            });
+        } else {
+            console.log('Operação de exclusão cancelada pelo usuário.');
+        }
     });
 });
